updated frontend to reflect db refactor
