+++ client/src/app/library/library.component.html
@@ -1,6 +1,6 @@
-<div class="container-fluid text-center my-4" *ngIf="mangas">
+<div class="container-fluid text-center my-4" *ngIf="mangaList">
   <div class="row row-cols-auto justify-content-center g-4">
-    <div class="col" *ngFor="let manga of mangas">
+    <div class="col" *ngFor="let manga of mangaList">
 
       <!-- <div class="card" style="width: 16rem; height: 24rem;"> -->
       <div class="card" style="width: 16rem;">
