+++ client/src/app/library/library.component.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { HttpErrorResponse } from '@angular/common/http';
 import { RouterLink } from '@angular/router';
 
-import { Manga } from '../interfaces';
+import { MinimalManga } from '../manga-models';
 import { BackendApiService } from '../backend-api.service';
 
 @Component({
@@ -14,7 +14,7 @@ import { BackendApiService } from '../backend-api.service';
     styleUrls: ['./library.component.css']
 })
 export class LibraryComponent implements OnInit {
-    mangas: Manga[] | undefined;
+    mangaList: MinimalManga[] | undefined;
 
     constructor(private apiService: BackendApiService) { }
 
@@ -24,18 +24,13 @@ export class LibraryComponent implements OnInit {
 
     private getManga(): void {
         this.apiService.getAllManga().subscribe({
-            next: (res: Manga[]) => {
-                this.mangas = res;
-                this.mangas.sort((m1, m2) => {
-                    if (m1.title > m2.title) return 1;
-                    if (m1.title < m2.title) return -1;
-                    return 0;
-                });
+            next: (res: MinimalManga[]) => {
+                this.mangaList = res;
             },
             error: (error: HttpErrorResponse) => {
                 console.error(error);
                 alert(error.message);
             }
-        })
+        });
     }
 }
