+++ client/src/app/manga-models.ts
@@ -1,5 +1,3 @@
-// TODO: update
-
 // Interface for new manga
 export interface NewManga {
     title: string;
@@ -13,7 +11,7 @@ export interface MinimalManga extends NewManga {
 }
 
 // Interface for manga in the detail and editing views
-export interface FullManga extends MinimalManga {
+export interface Manga extends MinimalManga {
     description: string;
     isMonitored: boolean;
     isDescriptionLocked: boolean;
@@ -23,30 +21,9 @@ export interface FullManga extends MinimalManga {
 }
 
 // Interface for chapters
-// export interface Chapter {
-//     slug: string;
-//     number: number;
-//     url: string;
-//     isRead: boolean;
-// }
-
-export interface Manga {
-    title: string;
-    slug: string;
-    url: string;
-    description?: string | null;
-    imageUrl?: string | null;
-    isMonitored?: boolean | null;
-    isDescriptionLocked?: boolean | null;
-    isImageUrlLocked?: boolean | null;
-}
-
 export interface Chapter {
-    id: string;
-    mangaSlug: string;
     slug: string;
     number: number;
-    name: string | null;
     url: string;
-    isRead: boolean | null;
+    isRead: boolean;
 }
