+++ client/src/app/sync.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpErrorResponse } from '@angular/common/http';
 
-import { Manga } from './interfaces';
+import { Manga } from './manga-models';
 import { BackendApiService } from './backend-api.service';
 
 @Injectable({
